build(frontend): switch over to vite, new Dart sass and more
create-react-app is deprecated (so does react-scripts; see
https://react.dev/blog/2025/02/14/sunsetting-create-react-app). Vite is
just an abritary choice (mainly I know how to work with it). The file
with JSX syntax are renamed so they have explicit jsx extension. The
explicit file extension in the import, `.js`, is removed. The image
imports are modified by adding `?url` to the end to tell vite that we
want the URL of the image.

Node Sass is replaced with the Dart one since it is also deprecated
(see https://sass-lang.com/blog/node-sass-is-end-of-life/). The
bootstrap import is fixed.

Since vite and these new packages require newer nodejs, nodejs version
is updated to 22.14.0. require() statements are replaced with the
import statements.