fix: plot comamnd
Integrate the plot command into the package's command line interface.

Drop the supprot for specifying the log folder for the plot. For a plot that is added on to the run simulation command, the log folder of that run will be used automatically. For a plot that is requested after, use has to specify which log folder to read from.

Switch to a dynamic version file.

Remove example/Main.py and example/MainPlot.py now that both run and plot commands are a bit stable.

Remove the hard-coded colormap lookup.